apps/api/requirements-mobile-agent.txt
livekit-agents[openai]~=1.0
python-dotenv>=1.0.1
